--- a/tools/golangci-override.yaml
+++ b/tools/golangci-override.yaml
@@ -4,3 +4,4 @@ linters-settings:
     - golang.org/x/net/http2/h2c: "h2c.NewHandler is unsafe; use wrapper istio.io/istio/pkg/h2c"
     - github.com/golang/protobuf/jsonpb: "don't use the jsonpb package directly; use util/protomarshal instead"
     - google.golang.org/protobuf/encoding/protojson: "don't use the protojson package directly; use util/protomarshal instead"
+    - gomodules.xyz/jsonpatch/v3: "don't use v3; v2 is orders of magnitude higher performance"