… go.sum
… licenses/gomodules.xyz/jsonpatch/v3/LICENSE
… licenses/gomodules.xyz/orderedmap/LICENSE