This picks up performance improvements I made into the library:

```
name                       old time/op    new time/op    delta
Injection/many_env_vars-8     2.13s ±18%     0.01s ± 8%  -99.48%  (p=0.008 n=5+5)

name                       old alloc/op   new alloc/op   delta
Injection/many_env_vars-8     908MB ± 0%       6MB ± 0%  -99.32%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
Injection/many_env_vars-8      749k ± 0%      106k ± 0%  -85.87%  (p=0.016 n=4+5)
```

Note we intentionally downgrade from v3 to v2. v3 is fundamentally slower due to an addition of stable ordering via "orderedmap". This is nice to have, but _tanks_ performance; we do not actually care about stability of the patch, only the output. v2 is guaranteed to result in stable results for _applying_ patches, just the raw patch ordering is undefined (as json in general has no ordering of keys). We used to have a test of the patch output, which has been removed -  we have over 100 test cases testing the full e2e path of applying the patch, so this test provided little value.