@istio-testing 2023-05-22
In response to a cherrypick label: #45029 failed to apply on top of branch "release-1.18":
```
Applying: Update jsonpatch library
Using index info to reconstruct a base tree...
M	common/config/.golangci.yml
M	go.mod
M	go.sum
M	pilot/pkg/config/kube/crdclient/client.go
Falling back to patching base and 3-way merge...
Auto-merging pilot/pkg/config/kube/crdclient/client.go
Auto-merging go.sum
Auto-merging go.mod
CONFLICT (content): Merge conflict in go.mod
Auto-merging common/config/.golangci.yml
error: Failed to merge in the changes.
hint: Use 'git am --show-current-patch=diff' to see the failed patch
Patch failed at 0001 Update jsonpatch library
When you have resolved this problem, run "git am --continue".
If you prefer to skip this patch, run "git am --skip" instead.
To restore the original branch and stop patching, run "git am --abort".

```