@istio-testing 2023-05-22
In response to a cherrypick label: new issue created for failed cherrypick: #45032